# platform/api-server/src/domains/correspondence/handlers/correspondence.handlers.ts
/**
 * Handlers for Correspondence
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: correspondence
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { CorrespondenceDomainModule } from "../dependencies/correspondence-ddd.dependencies.js";
import type { ListCaseCorrespondenceInput, SendCaseCorrespondenceInput } from "@forgetcase/services/correspondence";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/erasure-cases/{caseId}/correspondence
 * Operation: listCaseCorrespondence
 */
export async function listCaseCorrespondence(
  request: FastifyRequest<{ Params: { caseId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: CorrespondenceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListCaseCorrespondenceInput = {
    caseId: request.params.caseId,
  };
  const result = await deps.useCases.correspondences.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/erasure-cases/{caseId}/correspondence
 * Operation: sendCaseCorrespondence
 */
export async function sendCaseCorrespondence(
  request: FastifyRequest<{ Params: { caseId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: CorrespondenceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: SendCaseCorrespondenceInput = {
    caseId: request.params.caseId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.correspondences.create.execute(input);
  return reply.code(201).send(result);
}