# platform/api-server/src/domains/fulfilment/handlers/complete.handlers.ts
/**
 * Handlers for Complete
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: fulfilment
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { FulfilmentDomainModule } from "../dependencies/fulfilment-ddd.dependencies.js";
import type { CompleteBackupBeyondUseTaskInput, CompleteLiveErasureTaskInput } from "@forgetcase/services/fulfilment";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/live-erasure-tasks/{taskId}/complete
 * Operation: completeLiveErasureTask
 */
export async function completeLiveErasureTask(
  request: FastifyRequest<{ Params: { taskId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: FulfilmentDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CompleteLiveErasureTaskInput = {
    taskId: request.params.taskId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.completes.create.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/backup-beyond-use-tasks/{taskId}/complete
 * Operation: completeBackupBeyondUseTask
 */
export async function completeBackupBeyondUseTask(
  request: FastifyRequest<{ Params: { taskId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: FulfilmentDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CompleteBackupBeyondUseTaskInput = {
    taskId: request.params.taskId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.completes.create.execute(input);
  return reply.code(200).send(result);
}