# platform/api-server/src/domains/sla/handlers/extension.handlers.ts
/**
 * Handlers for Extension
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: sla
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { SlaDomainModule } from "../dependencies/sla-ddd.dependencies.js";
import type { ExtendSlaClockInput } from "@forgetcase/services/sla";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/erasure-cases/{caseId}/sla/extension
 * Operation: extendSlaClock
 */
export async function extendSlaClock(
  request: FastifyRequest<{ Params: { caseId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: SlaDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ExtendSlaClockInput = {
    caseId: request.params.caseId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.extensions.get.execute(input);
  return reply.code(200).send(result);
}